# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "This API provides endpoints for the various users of the Reward Loyalty script.",
        "contact": {
            "email": "support@nowsquare.com"
        },
        "version": "1.6.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "https://reward-loyalty-demo.nowsquare.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/partner/cards": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "Retrieve all active cards for the authenticated partner",
                "description": "Retrieve all active cards associated with the authenticated partner where the associated club is also active.",
                "operationId": "getPartnerCards",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cards retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Card"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/partner/cards/{cardId}": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "Retrieve a specific active card for the authenticated partner",
                "description": "Retrieve a specific active card by its ID associated with the authenticated partner.",
                "operationId": "getCard",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "cardId",
                        "in": "path",
                        "description": "Unique identifier of the card to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Response structure when user is unauthenticated",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "description": "Response structure when resource is not found",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Not found."
                    }
                },
                "type": "object"
            },
...
            "Card": {
                "required": [
                    "id",
                    "club_id",
                    "name",
                    "head",
                    "title",
                    "description",
                    "unique_identifier",
                    "issue_date",
                    "expiration_date",
                    "bg_color",
                    "bg_color_opacity",
                    "text_color",
                    "text_label_color",
                    "qr_color_light",
                    "qr_color_dark",
                    "currency",
                    "initial_bonus_points",
                    "points_expiration_months",
                    "currency_unit_amount",
                    "points_per_currency",
                    "min_points_per_purchase",
                    "max_points_per_purchase",
                    "is_visible_by_default",
                    "is_visible_when_logged_in",
                    "total_amount_purchased",
                    "number_of_points_issued",
                    "last_points_issued_at",
                    "number_of_points_redeemed",
                    "number_of_rewards_redeemed",
                    "last_reward_redeemed_at",
                    "views",
                    "last_view",
                    "created_by",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the Card",
                        "type": "integer",
                        "example": 81559402807296
                    },
                    "club_id": {
                        "description": "Unique identifier for the Club associated with the Card",
                        "type": "integer",
                        "example": 81559380115456
                    },
                    "name": {
                        "description": "Name of the Card",
                        "type": "string"
                    },
                    "head": {
                        "description": "Multilingual card name",
                        "properties": {
                            "de_DE": {
                                "type": "string",
                                "example": "Gesunde Ernährung"
                            },
                            "en_US": {
                                "type": "string",
                                "example": "Healthy Eats"
                            },
                            "es_ES": {
                                "type": "string",
                                "example": "Comida Saludable"
                            },
                            "fr_FR": {
                                "type": "string",
                                "example": "Repas Sains"
                            },
                            "nl_NL": {
                                "type": "string",
                                "example": "Gezond Eten"
                            },
                            "pt_BR": {
                                "type": "string",
                                "example": "Comidas Saudáveis"
                            }
                        },
                        "type": "object"
                    },
                    "title": {
                        "description": "Multilingual card title",
                        "properties": {
                            "de_DE": {
                                "type": "string",
                                "example": "Salat Ersparnisse"
                            },
                            "en_US": {
                                "type": "string",
                                "example": "Salad Savings"
                            },
                            "es_ES": {
                                "type": "string",
                                "example": "Ahorros en Ensaladas"
                            },
                            "fr_FR": {
                                "type": "string",
                                "example": "Économies sur les Salades"
                            },
                            "nl_NL": {
                                "type": "string",
                                "example": "Salade Besparingen"
                            },
                            "pt_BR": {
                                "type": "string",
                                "example": "Economia em Salada"
                            }
                        },
                        "type": "object"
                    },
                    "description": {
                        "description": "Multilingual card description",
                        "properties": {
                            "de_DE": {
                                "type": "string",
                                "example": "Belohnung für gesunde Entscheidungen!"
                            },
                            "en_US": {
                                "type": "string",
                                "example": "Get rewarded for healthy choices!"
                            },
                            "es_ES": {
                                "type": "string",
                                "example": "¡Obtén recompensas por elecciones saludables!"
                            },
                            "fr_FR": {
                                "type": "string",
                                "example": "Soignez vos choix et soyez récompensé(e) !"
                            },
                            "nl_NL": {
                                "type": "string",
                                "example": "Word beloond voor gezonde keuzes!"
                            },
                            "pt_BR": {
                                "type": "string",
                                "example": "Seja recompensado por escolhas saudáveis!"
                            }
                        },
                        "type": "object"
                    },
                    "unique_identifier": {
                        "description": "Unique identifier of the Card",
                        "type": "string"
                    },
                    "issue_date": {
                        "description": "Issue date of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "expiration_date": {
                        "description": "Expiration date of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "bg_color": {
                        "description": "Background color of the Card",
                        "type": "string",
                        "example": "#CCCCCC"
                    },
                    "bg_color_opacity": {
                        "description": "Background color opacity of the Card",
                        "type": "integer",
                        "example": 80
                    },
                    "text_color": {
                        "description": "Text color of the Card",
                        "type": "string",
                        "example": "#333333"
                    },
                    "text_label_color": {
                        "description": "Text label color of the Card",
                        "type": "string",
                        "example": "#333333"
                    },
                    "qr_color_light": {
                        "description": "QR code light color of the Card",
                        "type": "string",
                        "example": "#CCCCCC"
                    },
                    "qr_color_dark": {
                        "description": "QR code dark color of the Card",
                        "type": "string",
                        "example": "#222222"
                    },
                    "currency": {
                        "description": "Currency of the Card",
                        "type": "string",
                        "example": "USD"
                    },
                    "initial_bonus_points": {
                        "description": "Initial bonus points of the Card",
                        "type": "integer"
                    },
                    "points_expiration_months": {
                        "description": "Number of months after which the points on the Card expire",
                        "type": "integer"
                    },
                    "currency_unit_amount": {
                        "description": "Currency unit amount of the Card",
                        "type": "integer"
                    },
                    "points_per_currency": {
                        "description": "Number of points earned per currency unit spent using the Card",
                        "type": "integer"
                    },
                    "min_points_per_purchase": {
                        "description": "Minimum points that can be earned per purchase using the Card",
                        "type": "integer"
                    },
                    "max_points_per_purchase": {
                        "description": "Maximum points that can be earned per purchase using the Card",
                        "type": "integer"
                    },
                    "is_visible_by_default": {
                        "description": "Indicates if the Card is visible by default (1 for true, 0 for false)",
                        "type": "integer"
                    },
                    "is_visible_when_logged_in": {
                        "description": "Indicates if the Card is visible when a user is logged in (1 for true, 0 for false)",
                        "type": "integer"
                    },
                    "total_amount_purchased": {
                        "description": "Total amount purchased using the Card",
                        "type": "integer"
                    },
                    "number_of_points_issued": {
                        "description": "Total number of points issued for the Card",
                        "type": "integer"
                    },
                    "last_points_issued_at": {
                        "description": "Last time points were issued for the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "number_of_points_redeemed": {
                        "description": "Total number of points redeemed from the Card",
                        "type": "integer"
                    },
                    "number_of_rewards_redeemed": {
                        "description": "Total number of rewards redeemed using the Card",
                        "type": "integer"
                    },
                    "last_reward_redeemed_at": {
                        "description": "Last time a reward was redeemed using the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "views": {
                        "description": "Number of times the Card was viewed",
                        "type": "integer"
                    },
                    "last_view": {
                        "description": "Last time the Card was viewed",
                        "type": "string",
                        "format": "date-time"
                    },
                    "meta": {
                        "description": "Additional metadata associated with the Card in JSON format",
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "balance": {
                        "description": "The member's balance. This field has a value of 0 or higher for API calls where a member is authenticated. If the balance is not applicable or unavailable, its value is -1.",
                        "type": "integer",
                        "example": -1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "admin_auth_token": {
                "type": "http",
                "description": "Admin authentication token, formatted as `1|laravel_sanctum_<token_string>`.",
                "name": "Authorization",
                "scheme": "bearer"
            },
            "partner_auth_token": {
                "type": "http",
                "description": "Partner authentication token, formatted as `1|laravel_sanctum_<token_string>`.",
                "name": "Authorization",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Member authentication token, formatted as `1|laravel_sanctum_<token_string>`.",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}